uv/tests: fix warehouse lock file
This also appears to be a bug fix. The first change is to the resolution
markers, and apprently they were overlapping previously. Now they
aren't. The second change is to the marker for `setuptools`. Looking at
the dependency metadata for `ddtrace` confirms that this is correct:

```
$ curl -sL 'https://files.pythonhosted.org/packages/e2/88/c4a5044c69e929cfdeaf7d615a7f51d513ea74646ddaa5f5c17cb471e693/ddtrace-2.8.5-cp311-cp311-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata' | rg setuptools
Requires-Dist: setuptools ; python_version >= "3.12"
```